+++ src/main/java/gift/dto/orderDTO/OrderRequestDTO.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.orderDTO;
 
 public record OrderRequestDTO(
     Long optionId,
